Premium details update from new API

+++ app/src/main/java/bose/ankush/weatherify/domain/model/WeatherForecast.kt
@@ -13,10 +13,10 @@ data class WeatherForecast(
 ) {
     data class Alert(
         val description: String?,
-        val end: Int?,
+        val end: Long?,
         val event: String?,
         val sender_name: String?,
-        val start: Int?,
+        val start: Long?,
     )
 
     data class Current(
@@ -25,8 +25,8 @@ data class WeatherForecast(
         val feels_like: Double?,
         val humidity: Int?,
         val pressure: Int?,
-        val sunrise: Int?,
-        val sunset: Int?,
+        val sunrise: Long?,
+        val sunset: Long?,
         val temp: Double?,
         val uvi: Double?,
         val weather: List<WeatherCondition?>? = listOf(),
@@ -42,8 +42,8 @@ data class WeatherForecast(
         val pressure: Int?,
         val rain: Double?,
         val summary: String?,
-        val sunrise: Int?,
-        val sunset: Int?,
+        val sunrise: Long?,
+        val sunset: Long?,
         val temp: Temp?,
         val uvi: Double?,
         val weather: List<WeatherCondition?>? = listOf(),
